requirements.txt
selenium==4.15.2  
beautifulsoup4==4.12.2  
requests==2.31.0  
pandas==2.1.3  
openai==0.28.1  
webdriver-manager==4.0.1  
python-dotenv==1.0.0  
colorama==0.4.6  
tabulate==0.9.0  
html5lib==1.1
